# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "description": "Create a new user",
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Create User",
        "tags": [
          "Users Controller"
        ]
      },
      "get": {
        "description": "Get all users",
        "operationId": "UsersController_getAllUsers",
        "parameters": [
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset of users, default is 0",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit of users, default is 10",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllUsersResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get All Users",
        "tags": [
          "Users Controller"
        ]
      }
    },
    "/api/v1/users/short-uuid/{shortUuid}": {
      "get": {
        "description": "Get user by short UUID",
        "operationId": "UsersController_getUserByShortUuid",
        "parameters": [
          {
            "name": "shortUuid",
            "required": true,
            "in": "path",
            "description": "Short UUID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserByShortUuidResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get User By Short UUID",
        "tags": [
          "Users Controller"
        ]
      }
    },
    "/api/v1/users/sub-uuid/{subscriptionUuid}": {
      "get": {
        "description": "Get user by subscription UUID",
        "operationId": "UsersController_getUserBySubscriptionUuid",
        "parameters": [
          {
            "name": "subscriptionUuid",
            "required": true,
            "in": "path",
            "description": "UUID of the subscription",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserBySubscriptionUuidResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get User By Subscription UUID",
        "tags": [
          "Users Controller"
        ]
      }
    },
    "/api/v1/users/{uuid}": {
      "get": {
        "description": "Get user by UUID",
        "operationId": "UsersController_getUserByUuid",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "UUID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserByUuidResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Get User By UUID",
        "tags": [
          "Users Controller"
        ]
      }
    },
    "/api/v1/users/revoke/{uuid}": {
      "patch": {
        "description": "Revoke user subscription",
        "operationId": "UsersController_revokeUserSubscription",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "UUID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User subscription revoked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeUserSubscriptionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Revoke User Subscription",
        "tags": [
          "Users Controller"
        ]
      }
    },
    "/api/v1/users/disable/{uuid}": {
      "patch": {
        "description": "Disable user",
        "operationId": "UsersController_disableUser",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "UUID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User disabled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DisableUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Disable User",
        "tags": [
          "Users Controller"
        ]
      }
    },
    "/api/v1/users/enable/{uuid}": {
      "patch": {
        "description": "Enable user",
        "operationId": "UsersController_enableUser",
        "parameters": [
          {
            "name": "uuid",
            "required": true,
            "in": "path",
            "description": "UUID of the user",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User enabled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnableUserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "summary": "Enable User",
        "tags": [
          "Users Controller"
        ]
      }
    },
...
  },
  "info": {
    "title": "Remnawave API Schema",
    "description": "API for Remnawave",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT obtained login."
      }
    },
    "schemas": {
...
      "CreateUserRequestDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]+$",
            "maxLength": 32,
            "minLength": 6
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ACTIVE",
              "DISABLED",
              "LIMITED",
              "EXPIRED"
            ],
            "default": "ACTIVE"
          },
          "subscriptionUuid": {
            "type": "string",
            "format": "uuid"
          },
          "shortUuid": {
            "type": "string"
          },
          "trafficLimitBytes": {
            "description": "Traffic limit in bytes. 0 - unlimited",
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false
          },
          "trafficLimitStrategy": {
            "description": "Available reset periods",
            "type": "string",
            "enum": [
              "NO_RESET",
              "NO_RESET",
              "DAY",
              "WEEK",
              "MONTH",
              "YEAR"
            ],
            "default": "NO_RESET"
          },
          "enabledInbounds": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "uuid": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "required": [
                "uuid"
              ]
            }
          },
          "expireAt": {}
        },
        "required": [
          "username",
          "trafficLimitStrategy",
          "expireAt"
        ]
      },
      "CreateUserResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "uuid": {
                "type": "string",
                "format": "uuid"
              },
              "subscriptionUuid": {
                "type": "string",
                "format": "uuid"
              },
              "shortUuid": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "ACTIVE",
                  "DISABLED",
                  "LIMITED",
                  "EXPIRED"
                ],
                "default": "ACTIVE"
              },
              "usedTrafficBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitStrategy": {
                "description": "Available reset periods",
                "type": "string",
                "enum": [
                  "NO_RESET",
                  "NO_RESET",
                  "DAY",
                  "WEEK",
                  "MONTH",
                  "YEAR"
                ],
                "default": "NO_RESET"
              },
              "subLastUserAgent": {
                "type": "string",
                "nullable": true
              },
              "subLastIp": {
                "type": "string",
                "nullable": true
              },
              "expireAt": {},
              "onlineAt": {
                "nullable": true
              },
              "subRevokedAt": {
                "nullable": true
              },
              "createdAt": {},
              "updatedAt": {},
              "activeUserInbounds": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tag": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "uuid",
                    "tag"
                  ]
                }
              }
            },
            "required": [
              "uuid",
              "subscriptionUuid",
              "shortUuid",
              "username",
              "subLastUserAgent",
              "subLastIp",
              "expireAt",
              "onlineAt",
              "subRevokedAt",
              "createdAt",
              "updatedAt",
              "activeUserInbounds"
            ]
          }
        },
        "required": [
          "response"
        ]
      },
      "GetAllUsersResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "uuid": {
                  "type": "string",
                  "format": "uuid"
                },
                "subscriptionUuid": {
                  "type": "string",
                  "format": "uuid"
                },
                "shortUuid": {
                  "type": "string"
                },
                "username": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "ACTIVE",
                    "ACTIVE",
                    "DISABLED",
                    "LIMITED",
                    "EXPIRED"
                  ],
                  "default": "ACTIVE"
                },
                "usedTrafficBytes": {
                  "type": "integer",
                  "default": 0
                },
                "trafficLimitBytes": {
                  "type": "integer",
                  "default": 0
                },
                "trafficLimitStrategy": {
                  "description": "Available reset periods",
                  "type": "string",
                  "enum": [
                    "NO_RESET",
                    "NO_RESET",
                    "DAY",
                    "WEEK",
                    "MONTH",
                    "YEAR"
                  ],
                  "default": "NO_RESET"
                },
                "subLastUserAgent": {
                  "type": "string",
                  "nullable": true
                },
                "subLastIp": {
                  "type": "string",
                  "nullable": true
                },
                "expireAt": {},
                "onlineAt": {
                  "nullable": true
                },
                "subRevokedAt": {
                  "nullable": true
                },
                "createdAt": {},
                "updatedAt": {},
                "activeUserInbounds": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "uuid": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "tag": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "uuid",
                      "tag"
                    ]
                  }
                }
              },
              "required": [
                "uuid",
                "subscriptionUuid",
                "shortUuid",
                "username",
                "subLastUserAgent",
                "subLastIp",
                "expireAt",
                "onlineAt",
                "subRevokedAt",
                "createdAt",
                "updatedAt",
                "activeUserInbounds"
              ]
            }
          }
        },
        "required": [
          "response"
        ]
      },
      "GetUserByShortUuidResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "uuid": {
                "type": "string",
                "format": "uuid"
              },
              "subscriptionUuid": {
                "type": "string",
                "format": "uuid"
              },
              "shortUuid": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "ACTIVE",
                  "DISABLED",
                  "LIMITED",
                  "EXPIRED"
                ],
                "default": "ACTIVE"
              },
              "usedTrafficBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitStrategy": {
                "description": "Available reset periods",
                "type": "string",
                "enum": [
                  "NO_RESET",
                  "NO_RESET",
                  "DAY",
                  "WEEK",
                  "MONTH",
                  "YEAR"
                ],
                "default": "NO_RESET"
              },
              "subLastUserAgent": {
                "type": "string",
                "nullable": true
              },
              "subLastIp": {
                "type": "string",
                "nullable": true
              },
              "expireAt": {},
              "onlineAt": {
                "nullable": true
              },
              "subRevokedAt": {
                "nullable": true
              },
              "createdAt": {},
              "updatedAt": {},
              "activeUserInbounds": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tag": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "uuid",
                    "tag"
                  ]
                }
              }
            },
            "required": [
              "uuid",
              "subscriptionUuid",
              "shortUuid",
              "username",
              "subLastUserAgent",
              "subLastIp",
              "expireAt",
              "onlineAt",
              "subRevokedAt",
              "createdAt",
              "updatedAt",
              "activeUserInbounds"
            ]
          }
        },
        "required": [
          "response"
        ]
      },
      "GetUserBySubscriptionUuidResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "uuid": {
                "type": "string",
                "format": "uuid"
              },
              "subscriptionUuid": {
                "type": "string",
                "format": "uuid"
              },
              "shortUuid": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "ACTIVE",
                  "DISABLED",
                  "LIMITED",
                  "EXPIRED"
                ],
                "default": "ACTIVE"
              },
              "usedTrafficBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitStrategy": {
                "description": "Available reset periods",
                "type": "string",
                "enum": [
                  "NO_RESET",
                  "NO_RESET",
                  "DAY",
                  "WEEK",
                  "MONTH",
                  "YEAR"
                ],
                "default": "NO_RESET"
              },
              "subLastUserAgent": {
                "type": "string",
                "nullable": true
              },
              "subLastIp": {
                "type": "string",
                "nullable": true
              },
              "expireAt": {},
              "onlineAt": {
                "nullable": true
              },
              "subRevokedAt": {
                "nullable": true
              },
              "createdAt": {},
              "updatedAt": {},
              "activeUserInbounds": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tag": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "uuid",
                    "tag"
                  ]
                }
              }
            },
            "required": [
              "uuid",
              "subscriptionUuid",
              "shortUuid",
              "username",
              "subLastUserAgent",
              "subLastIp",
              "expireAt",
              "onlineAt",
              "subRevokedAt",
              "createdAt",
              "updatedAt",
              "activeUserInbounds"
            ]
          }
        },
        "required": [
          "response"
        ]
      },
      "GetUserByUuidResponseDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "uuid": {
                "type": "string",
                "format": "uuid"
              },
              "subscriptionUuid": {
                "type": "string",
                "format": "uuid"
              },
              "shortUuid": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ACTIVE",
                  "ACTIVE",
                  "DISABLED",
                  "LIMITED",
                  "EXPIRED"
                ],
                "default": "ACTIVE"
              },
              "usedTrafficBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitBytes": {
                "type": "integer",
                "default": 0
              },
              "trafficLimitStrategy": {
                "description": "Available reset periods",
                "type": "string",
                "enum": [
                  "NO_RESET",
                  "NO_RESET",
                  "DAY",
                  "WEEK",
                  "MONTH",
                  "YEAR"
                ],
                "default": "NO_RESET"
              },
              "subLastUserAgent": {
                "type": "string",
                "nullable": true
              },
              "subLastIp": {
                "type": "string",
                "nullable": true
              },
              "expireAt": {},
              "onlineAt": {
                "nullable": true
              },
              "subRevokedAt": {
                "nullable": true
              },
              "createdAt": {},
              "updatedAt": {},
              "activeUserInbounds": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "uuid": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "tag": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "uuid",
                    "tag"
                  ]
                }
              }
            },
            "required": [
              "uuid",
              "subscriptionUuid",
              "shortUuid",
              "username",
              "subLastUserAgent",
              "subLastIp",
              "expireAt",
              "onlineAt",
              "subRevokedAt",
              "createdAt",
              "updatedAt",
              "activeUserInbounds"
            ]
          }
        },
        "required": [
          "response"
        ]
      },
...
    }
  }
}